requirements.txt
fastapi==0.112.0
peewee==3.17.6
python-dotenv==1.0.1
pydantic~=2.8.2
pytest~=8.3.2
anyio~=4.4.0
PyJWT~=2.9.0
argon2-cffi